requirements.txt
GitPython==3.1.31
jinja2==3.1.2
MarkupSafe==2.1.2
Owlready2==0.40
Pillow==9.4.0
PyYAML==6.0
Pygments==2.14.0
Pympler==1.0.1
Send2Trash==1.8.0
altair==4.2.2
attrs=22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
experta==1.9.4
fonttools==4.39.0
frozendict==1.2
gitdb==4.0.10
idna=3.4
importlib-metadata==6.0.0
init==0.1.0
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.24.2
packaging=23.0
pandas==1.5.3
pip==23.0.1
plotly==5.13.1
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.2
schema==0.6.7
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
streamlit==1.20.0
tenacity==8.2.2
terminal==0.4.0
threadpoolctl==3.1.0
toml=0.10.2
toolz==0.12.0
tornado==6.2
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
webencodings==0.5.1
yellowbrick==1.5
zipp==3.15.0